# Improve performance of transform logic when many transforms are present