--- a/osu.Framework.Tests/Visual/Containers/TestSceneVisibilityContainer.cs
+++ b/osu.Framework.Tests/Visual/Containers/TestSceneVisibilityContainer.cs
@@ -1,6 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Linq;
 using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -125,7 +126,7 @@ namespace osu.Framework.Tests.Visual.Containers
 
         private class TestNestedVisibilityContainer : TestVisibilityContainer
         {
-            public bool BoxHasTransforms => box.Transforms.Count > 0;
+            public bool BoxHasTransforms => box.Transforms.Any();
 
             private readonly TestVisibilityContainer nested;
             private readonly Box box;