--- a/osu.Framework.Tests/Visual/Drawables/TestSceneTransformRewinding.cs
+++ b/osu.Framework.Tests/Visual/Drawables/TestSceneTransformRewinding.cs
@@ -53,7 +53,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(500, box => Precision.AlmostEquals(box.Scale.X, 0.5f));
             checkAtTime(250, box => Precision.AlmostEquals(box.Scale.X, 0.75f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 1);
+            AddAssert("check transform count", () => box.Transforms.Count() == 1);
         }
 
         [Test]
@@ -78,7 +78,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(interval * (i -= 2), box => Precision.AlmostEquals(box.Scale.X, 0.5f));
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.Scale.X, 0.75f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 4);
+            AddAssert("check transform count", () => box.Transforms.Count() == 4);
         }
 
         [Test]
@@ -105,7 +105,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(interval * (i -= 2), box => Precision.AlmostEquals(box.Y, 0.75f));
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.X, 0.75f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 4);
+            AddAssert("check transform count", () => box.Transforms.Count() == 4);
         }
 
         [Test]
@@ -131,7 +131,7 @@ namespace osu.Framework.Tests.Visual.Drawables
 
             checkAtTime(interval * (i - 2), box => Precision.AlmostEquals(box.Alpha, 1f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 7);
+            AddAssert("check transform count", () => box.Transforms.Count() == 7);
         }
 
         [Test]
@@ -154,7 +154,7 @@ namespace osu.Framework.Tests.Visual.Drawables
 
             checkAtTime(interval * ++i, box => Precision.AlmostEquals(box.Scale.X, 0.1f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 2);
+            AddAssert("check transform count", () => box.Transforms.Count() == 2);
         }
 
         [Test]
@@ -176,7 +176,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.Scale.X, 0.3125f));
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.Scale.X, 0.25f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 2);
+            AddAssert("check transform count", () => box.Transforms.Count() == 2);
         }
 
         [Test]
@@ -199,7 +199,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.Scale.X, 0.6875f));
             checkAtTime(interval * --i, box => Precision.AlmostEquals(box.Scale.X, 0.375f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 2);
+            AddAssert("check transform count", () => box.Transforms.Count() == 2);
         }
 
         [Test]
@@ -216,7 +216,7 @@ namespace osu.Framework.Tests.Visual.Drawables
             checkAtTime(interval * 4, box => Precision.AlmostEquals(box.Alpha, 1) && Precision.AlmostEquals(box.Scale.X, 0.9f));
             checkAtTime(interval * 2, box => Precision.AlmostEquals(box.Alpha, 0) && Precision.AlmostEquals(box.Scale.X, 0.575f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 3);
+            AddAssert("check transform count", () => box.Transforms.Count() == 3);
         }
 
         [Test]
@@ -238,8 +238,8 @@ namespace osu.Framework.Tests.Visual.Drawables
                     box.Delay(interval * 3).FadeOutFromOne(interval);
 
                     // FadeOutFromOne adds extra transforms which disallow testing this scenario, so we remove them.
-                    box.RemoveTransform(box.Transforms[2]);
-                    box.RemoveTransform(box.Transforms[0]);
+                    box.RemoveTransform(box.Transforms.ElementAt(2));
+                    box.RemoveTransform(box.Transforms.ElementAt(0));
                 }
             });
 
@@ -272,8 +272,8 @@ namespace osu.Framework.Tests.Visual.Drawables
                     box.Delay(interval * 3).FadeInFromZero(interval);
 
                     // FadeOutFromOne adds extra transforms which disallow testing this scenario, so we remove them.
-                    box.RemoveTransform(box.Transforms[2]);
-                    box.RemoveTransform(box.Transforms[0]);
+                    box.RemoveTransform(box.Transforms.ElementAt(2));
+                    box.RemoveTransform(box.Transforms.ElementAt(0));
                 }
             });
 
@@ -300,7 +300,7 @@ namespace osu.Framework.Tests.Visual.Drawables
                 checkAtTime(interval * i, box => Precision.AlmostEquals(box.Rotation, 0));
             }
 
-            AddAssert("check transform count", () => box.Transforms.Count == 10);
+            AddAssert("check transform count", () => box.Transforms.Count() == 10);
 
             for (int i = count; i >= 0; i--)
             {
@@ -316,7 +316,7 @@ namespace osu.Framework.Tests.Visual.Drawables
 
             checkAtTime(750, box => Precision.AlmostEquals(box.Rotation, 0f));
 
-            AddAssert("check transform count", () => box.Transforms.Count == 8);
+            AddAssert("check transform count", () => box.Transforms.Count() == 8);
 
             const int count = 4;
 
@@ -326,7 +326,7 @@ namespace osu.Framework.Tests.Visual.Drawables
                 checkAtTime(interval * i, box => Precision.AlmostEquals(box.Rotation, 0));
             }
 
-            AddAssert("check transform count", () => box.Transforms.Count == 10);
+            AddAssert("check transform count", () => box.Transforms.Count() == 10);
 
             for (int i = count; i >= 0; i--)
             {