--- a/osu.Framework/Graphics/Transforms/TargetMemberTransformTracker.cs
+++ b/osu.Framework/Graphics/Transforms/TargetMemberTransformTracker.cs
@@ -260,13 +260,13 @@ namespace osu.Framework.Graphics.Transforms
 
             invokePendingRemovalActions();
 
+            var time = transformable.Time.Current;
+
             // If our newly added transform could have an immediate effect, then let's
             // make this effect happen immediately.
-            if (transform.StartTime < Time.Current || transform.EndTime <= Time.Current)
-                UpdateTransforms(Time.Current, !transformable.RemoveCompletedTransforms && transform.StartTime <= Time.Current);
+            if (transform.StartTime < time || transform.EndTime <= time)
+                UpdateTransforms(time, !transformable.RemoveCompletedTransforms && transform.StartTime <= time);
         }
-
-        public FrameTimeInfo Time => transformable.Time;
 
         /// <summary>
         /// Removes <see cref="Transform"/>s that start after <paramref name="time"/>.