Focuses on decreasing the iteration overhead of per-frame transform processing when many transforms are present in the transforms list (and cannot be removed due to rewind support or otherwise).

# Basic idea

Store the last fully applied transform index and only continue processing from that point onwards. To keep things simple, invalidate this cache on any add/remove/rewind operation. These scenarios can be optimised in the future if/when required.

# Difficulties

To make this work, transforms had to be split into per-`TargetMember` lists. If this was not the case, it becomes hard to choose a starting index (originally I was using the "minimum" of any `TargetMember` but this can lead to situations with no optimisation is possible).

As such, there is now one new `TargetMemberTransformTracker` class per `TargetMember` being tracked for a `Transformable`. These are only constructed when used and the memory / allocation overhead does not seem any worse than it was previously.

# Benchmarks

## UpdateTransformsWithManyPresent 

|                          Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| (master)  | 821.0 ms | 22.08 ms | 65.09 ms |     - |     - |     - |      48 B |
| (this PR) | 0.2574 ms | 3.84 us  | 3.40 us  |     - |     - |     - |       2 B |

- Closes https://github.com/ppy/osu/issues/9605

- [x] Add benchmarks (not sure if required; osu!-side storyboard shows an improvement from 40ms frame times to <0.1ms)
- [x] Add lazy initialisation of dictionary (similar to `transformsLazy`)
- [x] Assess whether dictionary should only be initialised/used over a certain list length (to save on memory @smoogipoo thoughts?)
- [x] Move transform add / remove / removeAt into dedicated methods to avoid needing to do the cache clear at each usage?

I've implemented this in the simplest way possible. Further optimisations (adjusting indices rather than completing flushing the cache) are obviously possible but come with added complexity. Mainly PRing early for feedback (cases which may need to be tested and aren't already covered; general direction etc.) and considerations.