@peppy 2020-07-31
For verbosity, the test failure is from a very specific case as demonstrated by the test:

```
i=0  t=100 --------------- t = 1000
i=1       t=500  ------ t=750
```

Because I was using the `Max` last applied time, i=0 would be skipped in some cases.

I've fixed this issue by using `Min` and tracking `Applied` state as well.

As an aside, if this change is considered too complex / illegible in `Transformable`, the direction (which I believe we have discussed before) would be to split out transform logic into one processor class per TargetMember, maintaining its own pure list.

Of course, there are extra considerations if we go in this direction (ie. how do we apply transforms in chronological sequence, rather than per-member.. if this is even required) which I have not properly thought over.