@peppy 2020-07-31
Blocking for now, will re-visit with per-TargetMember implementation as the fix for that test failure has regressed the effectiveness of the optimisation. While still possible, it'd likely require two dictionaries in Transformable (tracking `Applied` and `AppliedtoEnd`) at which point the complexity seems too high.