@peppy 2020-08-03
Good to go.

Before:
```
|                          Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|
| UpdateTransformsWithManyPresent | 821.0 ms | 22.08 ms | 65.09 ms |     - |     - |     - |      48 B |
```

After:
```
|                          Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
| UpdateTransformsWithManyPresent | 257.4 us | 3.84 us | 3.40 us |     - |     - |     - |       2 B |
```